// .github/sync-issues/lib/types.ts
export interface RepoInfo {
    readonly owner: string;
    readonly repo: string;
}

/** GitHub Issue data */
export interface IssueData {
    readonly number: number;
    readonly id: string;
    readonly nodeId: string;
    readonly url: string;
    readonly title: string;
}
// ...
export type IssueTypeName = "Epic" | "Feature" | "Flag";

/** Issue creation request */
export interface CreateIssueRequest {
    readonly owner: string;
    readonly repo: string;
    readonly title: string;
    readonly body: string;
    readonly type?: IssueTypeName;
}